Conflicts:
	.gitattributes
	src/gov/nasa/jpl/ae/event/Call.java
	src/gov/nasa/jpl/ae/event/ConstructorCall.java
	src/gov/nasa/jpl/ae/event/Consumable.java
	src/gov/nasa/jpl/ae/event/EventSimulation.java
	src/gov/nasa/jpl/ae/event/Expression.java
	src/gov/nasa/jpl/ae/event/FunctionCall.java
	src/gov/nasa/jpl/ae/event/Functions.java
	src/gov/nasa/jpl/ae/event/Parameter.java
	src/gov/nasa/jpl/ae/event/TimeVaryingMap.java
	src/gov/nasa/jpl/ae/magicdrawPlugin/AE/main.py
	src/gov/nasa/jpl/ae/solver/CollectionTree.java
	src/gov/nasa/jpl/ae/util/ClassUtils.java
	src/gov/nasa/jpl/ae/util/CompareUtils.java
	src/gov/nasa/jpl/ae/util/Debug.java
	src/gov/nasa/jpl/ae/util/MoreToString.java
	src/gov/nasa/jpl/ae/util/Pair.java
	src/gov/nasa/jpl/ae/util/Timer.java
	src/gov/nasa/jpl/ae/util/Utils.java
	src/gov/nasa/jpl/ae/xml/EventXmlToJava.java
	src/gov/nasa/jpl/ae/xml/JavaForFunctionCall.java
	src/gov/nasa/jpl/ae/xml/XmlUtils.java

+++ src/gov/nasa/jpl/ae/solver/CollectionTree.java
@@ -1,11 +1,10 @@
 package gov.nasa.jpl.ae.solver;
 
-import gov.nasa.jpl.ae.util.ClassUtils;
-import gov.nasa.jpl.ae.util.MoreToString;
-import gov.nasa.jpl.ae.util.Pair;
-import gov.nasa.jpl.ae.util.Utils;
+import gov.nasa.jpl.mbee.util.ClassUtils;
+import gov.nasa.jpl.mbee.util.MoreToString;
+import gov.nasa.jpl.mbee.util.Pair;
+import gov.nasa.jpl.mbee.util.Utils;
 
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.AbstractCollection;
@@ -688,12 +687,12 @@ import java.util.TreeMap;
       HashSet<Object> h = new HashSet<Object>();
       ct.add(h);
       h.add( "e" );
-      h.add( 2.71 );
+      h.add( (Double)2.71 );
       h.add( (double)9.8 );
       TreeMap<String,Object> tm = new TreeMap< String, Object >();
       h.add( tm );
       tm.put( "foo", "bar" );
-      tm.put( "gravity", new Float(9.8) );
+      tm.put( "gravity", new Double(9.8) );
       java.util.Iterator i = ct.iterator();
       System.out.println("ct = " + ct );
     }
